requirements.txt
tiktoken
openai
streamlit
dotenv
pydantic-settings
typer